move yaml parsing into its own file and corrected incorrect delays
